requirements.txt
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
flask==3.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1